deploy: 7b864fb27258e089036a2182339d9c29f5066762

+++ bootstrap.sh
@@ -1,9 +1,8 @@
 #!/bin/sh
-rm .bashrc .bash_profile
-rm -r .vim
-rm -r .config/nvim
+[ -d .vim ] && rm -r .vim
+[ -d .config/nvim ] && rm -r .config/nvim
 git init
 git remote add origin git@github.com:renzmann/renzmann
 git pull
-git checkout main
+git checkout --force main
 git submodule update --init --recursive
